+++ tests/config/mod.rs
@@ -4,5 +4,9 @@
 pub mod compiler_artifact_parser_tests;
 pub mod job_list_builder_tests;
 pub mod job_source_tests;
+pub mod module_path_tests;
 pub mod pre_build_resolver_tests;
 pub mod pre_build_runner_tests;
+pub mod rust_file_finder_tests;
+pub mod test_name_reader_tests;
+pub mod test_path_scanner_tests;